Moved ProgressBar from gulpfile.js to tools/progress-bar.js.
